+++ public/index.php
@@ -1,11 +1,13 @@
 <?php
+    // This is the path to initialize.php, your site's gateway to the rest of the UF codebase!  Make sure that it is correct!
+    $init_path = "../userfrosting/initialize.php";
+
     // This if-block just checks that the path for initialize.php is correct.  Remove this once you know what you're doing.
-    if (!file_exists("../userfrosting/initialize.php")){
+    if (!file_exists($init_path)){
         echo "<h2>We can't seem to find our way to initialize.php!  Please check the require_once statement at the top of index.php, and make sure it contains the correct path to initialize.php.</h2><br>";
     }
 
-    // This is the path to initialize.php, your site's gateway to the rest of the UF codebase!  Make sure that it is correct!
-    require_once("../userfrosting/initialize.php");
+    require_once($init_path);
 
     use UserFrosting as UF;
    
@@ -50,7 +52,8 @@
         }
     })->name('uri_home');
 
-    // Miscellaneous pages
+    /********** FEATURE PAGES **********/
+    
     $app->get('/dashboard/?', function () use ($app) {    
         // Access-controlled page
         error_log("Checking access");
@@ -70,7 +73,8 @@
         $app->render('users/zerg.twig'); 
     });    
        
-    // Account management pages
+    /********** ACCOUNT MANAGEMENT INTERFACE **********/
+    
     $app->get('/account/:action/?', function ($action) use ($app) {    
         // Forward to installation if not complete
         if (!isset($app->site->install_status) || $app->site->install_status == "pending"){
@@ -112,12 +116,15 @@
         }
     });    
     
-    // User management pages
+    /********** USER MANAGEMENT INTERFACE **********/
+    
+    // List users
     $app->get('/users/?', function () use ($app) {
         $controller = new UF\UserController($app);
         return $controller->pageUsers();
     });    
 
+    // List users in a particular primary group
     $app->get('/users/:primary_group/?', function ($primary_group) use ($app) {
         $controller = new UF\UserController($app);
         return $controller->pageUsers($primary_group);
@@ -163,11 +170,19 @@
         return $controller->deleteUser($user_id);
     });
     
-    // Group management pages
+    /********** GROUP MANAGEMENT INTERFACE **********/
+    
+    // List groups
     $app->get('/groups/?', function () use ($app) {
         $controller = new UF\GroupController($app);
         return $controller->pageGroups();
     }); 
+    
+    // List auth rules for a group
+    $app->get('/groups/g/:group_id/auth?', function ($group_id) use ($app) {
+        $controller = new UF\GroupController($app);
+        return $controller->pageGroupAuthorization($group_id);
+    })->name('uri_authorization');  
     
     // Group info form (update/view)
     $app->get('/forms/groups/g/:group_id/?', function ($group_id) use ($app) {
@@ -203,11 +218,46 @@
         return $controller->deleteGroup($group_id);
     });
     
-    // Admin tools
+    /********** GROUP AUTH RULES INTERFACE **********/
+    
+    // Group auth creation form
+    $app->get('/forms/groups/g/:group_id/auth/?', function ($group_id) use ($app) {
+        $controller = new UF\AuthController($app);
+        return $controller->formAuthCreate($group_id, "group");
+    });      
+    
+    // Group auth update form
+    $app->get('/forms/groups/auth/a/:rule_id/?', function ($rule_id) use ($app) {
+        $controller = new UF\AuthController($app);
+        $get = $app->request->get();        
+        return $controller->formAuthEdit($rule_id);
+    });    
+
+    // Group auth create
+    $app->post('/groups/g/:group_id/auth/?', function ($group_id) use ($app) {
+        $controller = new UF\AuthController($app);
+        return $controller->createAuthRule($group_id, "group");
+    });     
+
+    // Group auth update
+    $app->post('/groups/auth/a/:rule_id?', function ($rule_id) use ($app) {
+        $controller = new UF\AuthController($app);
+        return $controller->updateAuthRule($rule_id, "group");
+    });
+    
+    // Group auth delete
+    $app->post('/auth/a/:rule_id/delete/?', function ($rule_id) use ($app) {
+        $controller = new UF\AuthController($app);
+        $get = $app->request->get();        
+        return $controller->deleteAuthRule($rule_id);
+    });  
+        
+    /************ ADMIN TOOLS *************/
+    
     $app->get('/config/settings/?', function () use ($app) {
         $controller = new UF\AdminController($app);
         return $controller->pageSiteSettings();
-    })->name('uri_settings');   
+    })->name('uri_settings');     
     
     $app->post('/config/settings/?', function () use ($app) {
         $controller = new UF\AdminController($app);
@@ -226,7 +276,42 @@
         $app->redirect($app->urlFor('uri_settings'));
     });    
     
-    // Installation pages
+    // Slim info page
+    $app->get('/sliminfo/?', function () use ($app) {
+        // Access-controlled page
+        if (!$app->user->checkAccess('uri_slim_info')){
+            $app->notFound();
+        }
+        echo "<pre>";
+        print_r($app->environment());
+        echo "</pre>";
+    });
+
+    // PHP info page
+    $app->get('/phpinfo/?', function () use ($app) {
+        // Access-controlled page
+        if (!$app->user->checkAccess('uri_php_info')){
+            $app->notFound();
+        }    
+        echo "<pre>";
+        print_r(phpinfo());
+        echo "</pre>";
+    });
+
+    // Error log page
+    $app->get('/errorlog/?', function () use ($app) {
+        // Access-controlled page
+        if (!$app->user->checkAccess('uri_error_log')){
+            $app->notFound();
+        }
+        $log = $app->site->getLog();
+        echo "<pre>";
+        echo implode("<br>",$log['messages']);
+        echo "</pre>";
+    });      
+       
+    /************ INSTALLER *************/
+
     $app->get('/install/?', function () use ($app) {
         $controller = new UF\InstallController($app);
         if (isset($app->site->install_status)){
@@ -259,45 +344,15 @@
         }   
     });
     
-    // API
+    /************ API *************/
+    
     $app->get('/api/users/?', function () use ($app) {
         $controller = new UF\ApiController($app);
         $controller->listUsers();
     });
     
-    // Slim info page
-    $app->get('/sliminfo/?', function () use ($app) {
-        // Access-controlled page
-        if (!$app->user->checkAccess('uri_slim_info')){
-            $app->notFound();
-        }
-        echo "<pre>";
-        print_r($app->environment());
-        echo "</pre>";
-    });
-
-    // PHP info page
-    $app->get('/phpinfo/?', function () use ($app) {
-        // Access-controlled page
-        if (!$app->user->checkAccess('uri_php_info')){
-            $app->notFound();
-        }    
-        echo "<pre>";
-        print_r(phpinfo());
-        echo "</pre>";
-    });
-
-    // PHP info page
-    $app->get('/errorlog/?', function () use ($app) {
-        // Access-controlled page
-        if (!$app->user->checkAccess('uri_error_log')){
-            $app->notFound();
-        }
-        $log = $app->site->getLog();
-        echo "<pre>";
-        echo implode("<br>",$log['messages']);
-        echo "</pre>";
-    });      
+    
+    /************ MISCELLANEOUS UTILITY ROUTES *************/
     
     // Generic confirmation dialog
     $app->get('/forms/confirm/?', function () use ($app) {
@@ -353,31 +408,6 @@
             $controller->page404();
         } else {
             $app->alerts->addMessageTranslated("danger", "SERVER_ERROR");
-        }
-    });     
-    
-    $app->get('/test/auth', function() use ($app){
-        if (0 == "php")
-            echo "0 = php";
-        
-        $params = [
-            "user" => [
-                "id" => 1
-            ],
-            "post" => [
-                "id" => 7
-            ]
-        ];
-        
-        $conditions = "(equals(self.id,user.id)||hasPost(self.id,post.id))&&subset(post, [\"id\", \"title\", \"content\", \"subject\", 3])";
-        
-        $ace = new UF\AccessConditionExpression($app);
-        $result = $ace->evaluateCondition($conditions, $params);
-        
-        if ($result){
-            echo "Passed $conditions";
-        } else {
-            echo "Failed $conditions";
         }
     });