+++ public/js/widget-users.js
@@ -1,29 +1,10 @@
-/*
-
-UserFrosting
-By Alex Weissman
-
-UserFrosting is 100% free and open-source.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the 'Software'), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED 'AS IS', WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-
-*/
-
+/**
+ * @file This file contains functions and bindings for the UserFrosting user management pages.
+ *
+ * @author Alex Weissman
+ * @license MIT
+ */
+ 
 $(document).ready(function() {                   
     bindUserTableButtons($("body"));
 });
@@ -175,7 +156,9 @@ function deleteUserDialog(box_id, user_id, name){
 	});
 }
 
-/* Display a modal form for updating/creating a user */
+/**
+ * Display a modal form for updating/creating a user.
+ */
 function userForm(box_id, user_id) {	
 	user_id = typeof user_id !== 'undefined' ? user_id : "";